+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 ### Fixed
 
 - Fixed incorrectly calculated module descriptor coordinates used for excluding transitive dependencies. JetBrains/intellij-platform-gradle-plugin#2127 
+- Refactor `ComposedJarTask` module naming logic and remove redundant test method from `IntelliJPlatformTestingExtensionTest`. JetBrains/intellij-platform-gradle-plugin#2124 
 
 ## [2.14.0] - 2026-04-09